// frontend/src/api/nominalMetrics.ts
export interface NominalMetrics {
  id: number;
  exchange: string;
  ticker: string;
  currency: string;
  favorite: boolean;
  industry: string;
  nextEarningsDate: Date | undefined;
  revenueTtm: number | undefined;
  revenueGrowthYoyTtm: number | undefined;
  revenueGrowthMultiYear: number | undefined;
  grossProfitGrowthYoyTtm: number | undefined;
  grossProfitGrowthMultiYear: number | undefined;
  grossMarginTtm: number | undefined;
  grossMarginShortTermTrend: string | undefined;
  grossMarginLongTermTrend: string | undefined;
  sgaRatioTtm: number | undefined;
  sgaShortTermTrend: string | undefined;
  sgaLongTermTrend: string | undefined;
  rndRatioTtm: number | undefined;
  rndShortTermTrend: string | undefined;
  rndLongTermTrend: string | undefined;
  operatingMarginTtm: number | undefined;
  operatingMarginShortTermTrend: string | undefined;
  operatingMarginLongTermTrend: string | undefined;
  interestExpenseRatioTtm: number | undefined;
  netMarginTtm: number | undefined;
  theoreticalNetMargin: number | undefined;
  isNetMarginOptimized: boolean | undefined;
  sharesOutstandingDilutedTtm: number | undefined;
  sharesChangeTtm: number | undefined;
  sharesChangeMultiYear: number | undefined;
  sharesChangeTrend: string | undefined;
  retainedEarningsTtm: number | undefined;
  retainedEarningsTrend: string | undefined;
  netCashTtm: number | undefined;
  hasHealthyNetCash: boolean | undefined;
  netCashTrend: string | undefined;
  operatingCashFlowTtm: number | undefined;
  operatingCashFlowMarginTtm: number | undefined;
  operatingCashFlowMarginTrend: string | undefined;
  freeCashFlowTtm: number | undefined;
  freeCashFlowMarginTtm: number | undefined;
  revenueNextYear: number | undefined;
  revenueGrowthNextYear: number | undefined;
  priceCurrentRevenueGrowth: number | undefined;
  priceCurrentGpGrowth: number | undefined;
  priceNextYearRevenueGrowth: number | undefined;
  priceMultiYearRevenueGrowth: number | undefined;
  priceMultiYearGpGrowth: number | undefined;
  latestPrice: number | undefined;
  lastUpdated: Date | undefined;
}